Extract class Pull::Diff from Pull::Show
Diff is an own subcommand now.
